separated follow button to its own compoennt so it doesn't rerender everything. Will clean up code after demo
