+++ front-end/src/pages/ProfilePage.js
@@ -1,17 +1,14 @@
 import React, { useContext, useState, useEffect } from "react";
 import { useNavigate, Link } from "react-router-dom";
 import { DarkModeContext } from "../context/DarkModeContext";
-import { useAuthContext } from "../hooks/useAuthContext";
+import Follow from "../components/Follow";
 
 function ProfilePage() {
   const navigate = useNavigate();
-  const { user } = useAuthContext();
   const [fakeName, setFakeName] = useState(""); // remove after sprint 1, only used to randomize displayed username using mockaroo
   const [description, setDescription] = useState("");
   const [postError, setPostError] = useState(null);
   const [loading, setLoading] = useState(true);
-  const [followed, setFollowed] = useState(false);
-  const [followedChanged, setFollowedChanged] = useState(false);
   const { ifDarkMode } = useContext(DarkModeContext);
 
   const randomSize = [350, 300, 250, 200, 230, 240, 310, 320, 330, 360, 380];
@@ -22,7 +19,6 @@ function ProfilePage() {
   const randomIndex5 = Math.floor(Math.random() * randomSize.length);
 
   useEffect(() => {
-    setFollowedChanged(false);
     async function fetchProfileInfo() {
         const response = await fetch(`https://my.api.mockaroo.com/post.json?key=9e339cc0`, {
             method: "GET",
@@ -46,41 +42,13 @@ function ProfilePage() {
             setLoading(false);
         }
     }
-    console.log(user);
     fetchProfileInfo();
-  }, [followedChanged]);
+  }, []);
 
-  const handleFollow = () => {
-    setFollowed(!followed);
-    setFollowedChanged(true);
-  };
-
-  function FollowButton() {
-    if (!followed){
-      return (
-        <button
-          onClick={() => handleFollow()}
-          className='btn btn-secondary'
-        >
-            Follow
-        </button>
-      )
-    }
-    else{
-      return (
-        <button
-            onClick={() => handleFollow()}
-            className='btn btn-primary'
-        >
-            Followed
-        </button>
-      )
-    }
-  }
   
   return (
     <div className={`profileContainer ${ifDarkMode && "darkTheme"}`}>
-      <div className="right">{user ? <FollowButton /> : <div></div>}</div>
+      <Follow/>
       <div className="pfpContainer">
         <div className="pfp">
           <img src={`https://picsum.photos/${randomSize[randomIndex5]}/300`} alt="pic" />
